% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSEParser.R
\name{SSEparser}
\alias{SSEparser}
\title{Parse a Server Sent Event}
\description{
This class can help you parse a single server sent event or a stream of them.
You can inherit the class for a custom aplication.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{List  that contains all the events parsed. When the class is initialized, is just an empty list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SSEparser-append_parsed_sse}{\code{SSEparser$append_parsed_sse()}}
\item \href{#method-SSEparser-parse_sse}{\code{SSEparser$parse_sse()}}
\item \href{#method-SSEparser-new}{\code{SSEparser$new()}}
\item \href{#method-SSEparser-clone}{\code{SSEparser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSEparser-append_parsed_sse"></a>}}
\if{latex}{\out{\hypertarget{method-SSEparser-append_parsed_sse}{}}}
\subsection{Method \code{append_parsed_sse()}}{
Takes a parsed event and appends it to the \code{events} field. You can overwrite this method if you decide to extend this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSEparser$append_parsed_sse(parsed_event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_event}}{Event to append to the \code{events} field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSEparser-parse_sse"></a>}}
\if{latex}{\out{\hypertarget{method-SSEparser-parse_sse}{}}}
\subsection{Method \code{parse_sse()}}{
Takes a string that comes from a server sent event and parses it to an R list. You should never overwrite this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSEparser$parse_sse(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A server sent event as specified in the \href{https://html.spec.whatwg.org/multipage/server-sent-events.html#server-sent-events}{HTML spec}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSEparser-new"></a>}}
\if{latex}{\out{\hypertarget{method-SSEparser-new}{}}}
\subsection{Method \code{new()}}{
Create a new SSE parser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSEparser$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSEparser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SSEparser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSEparser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
